Use V7 to generate DKG request UUID (#36)
Upgraded UUID handling in DKG coordinator and added key index derivation functionality.

Since this is a distributed key generation, it's hard to communicate the generated key id to each individual participants. So use a deterministic way to generate key IDs.

The reason to use V7 is that V7 has a timestamp in the UUID that could avoid collision as much as possible