Adds challenge-based login (#187)
Adds a challenge-based login flow for spark by signing a challenge.

Overall, the flow is that the user will request a challenge from the
server for his public key, then sign the challenge and return it to the
server via the verify_challenge endpoint. The server will validate that
the challenge itself is not expired and that the key did indeed sign the
challenge correctly.

Currently there is no server-side session storage, so there is no way to
revoke sessions or really log out other than just deleting the session
token. Real sessions can come later =)


The next change will be to receive the token from requests (from
header?) and parse it into a ViewerContext-like object

Tests pass.