golint -> golangci-lint (#203)
This also enables a few other linters besides gofumpt and revive.

name | description 
------| --------
errcheck | Errcheck is a program for checking for unchecked errors in Go
code. These unchecked errors can be critical bugs in some cases.
gosimple | Linter for Go source code that specializes in simplifying
code.
govet | Vet examines Go source code and reports suspicious constructs.
It is roughly the same as 'go vet' and uses its passes.
ineffassign | Detects when assignments to existing variables are not
used.
staticcheck | It's a set of rules from staticcheck. It's not the same
thing as the staticcheck binary. The author of staticcheck doesn't
support or approve the use of staticcheck as a library inside
golangci-lint.
unused | Checks Go code for unused constants, variables, functions and
types.
gofumpt | Checks if code and import statements are formatted, with
additional rules.
revive | Fast, configurable, extensible, flexible, and beautiful linter
for Go. Drop-in replacement of golint.

Lints will ignore any existing errors as of 4463971 and we can just fix
forward. I have tested this functionality out locally!