Verify bitcoin script after signing (#215)
Our signatures might be right, but we also need to make sure they
validate when we construct the witness/transactions properly.

The main change to transaction construction is that we should not be
populating the scriptSig field on inputs since everything is a segwit
tx, i.e. signature data is all in the witness.

Everything seems to work in the grpc test with this, except for the
aggregate tree test.