Add session to context via AuthnInterceptor; integrate with StartSendTransfer (#193)
Reads session token data from the authorization metadata field of the
form `Bearer <token>`.

Not sure exactly what the product requirements are for session
validation for what particular APIs, but I've started with
StartSendTransfer. Once this has been approved, it will be pretty
trivial to add to the rest of the APIs.