Implement coop exit flow part 3 (#233)
Send the leaf key tweak and check the transfer is pending. To do this,
we make sure to sync the DB state between operators, by adding a new
method on the internal server proto, i.e. initiate cooperative exit.

All operators should check that the exit txid shows up on-chain before
finalizing transfers (this will be implemented and tested in an upcoming
PR).