Implement leaf swap flow (#246)
Implements leaf swap flow using adaptor signatures.

Outline
- Alice signs refunds for bob
- Alice derives adaptor sig/pubkey and gives to bob
- Bob signs refunds for alice using adaptor public key and gives to
alice
- Alice reveals adaptor secret to bob (bob can now form valid signatures
and we check this in the test) and alice provides key tweak to SO
- Bob claims alice's leaves
- Bob provides key tweak to SO
- Alice claims bob's leaves

This adds one new RPC call so that bob can ask the SOs to sign refunds
with the adaptor public key. It's basically just the start transfer call
but with an adaptor public key. Will probably be merged into that single
call in the future.