Ensure depositing funds after tree creation works (#249)
In [this
thread](https://lightsparkgroup.slack.com/archives/C07MP81322E/p1738353459957799)
we discussed a trusted deposit where the user sends funds to an address
between them and an SE before getting refund signatures.

I realized in our code, this is actually the default flow, and the
trustless flow didn't actually work. This PR checks to update a pending
tree and tree leaves if the L1 tx for that tree has confirmed.