Token Transaction Part 27: Add integration test for duplicate starts and add timestamp to mint operation (#457)
***TLDR***
-
Add a test to ensure that token transactions that are started but not signed can be started and signed by a different call for both issuance and transfer.

This change necessitated adding a nonce to the MintTokenInput to prevent a unique key violation on the issuer signature.  Use a timestamp for this to also give us a bit of extra data about when the issuer exactly intended to issue which could come in handy later.