Setup basic DB migrations (#367)
I chose the tool atlas because that's what the [golang ent docs
suggest](https://entgo.io/docs/versioned-migrations), and seems to work
pretty well out of the box.

This commits the initial DB baseline migration files, from which further
diffs will be created. This adds instructions for creating and applying
DB migrations. And also adds a migration step to the run-everything.sh.

We should apply migrations to the dev DB in CI, will probably need
@mgorven's help with that.