Add refresh timelock RPC (#502)
When a user hits the end of their timelock, they may want to reset their
refund tx timelock, and decrement the ancestor so they can continue to
send these funds. We implement a new RPC call to allow users to do this
generically for any number of ancestors.

There is a flake in the tests right now, but it passes most of the time.