[web-demo] this change improves the get all transactions query behavior and adds infinite scrolling to the transactions tab. added some styling changes to support this feature (#850)
CNT-151

[Screen Recording 2025-03-06 at 8.12.21 PM.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/AMCzR6ldnuRamjbkxWC9/cf8fab2d-9a83-4867-b43d-33e529256f10.mov" />](https://app.graphite.dev/media/video/AMCzR6ldnuRamjbkxWC9/cf8fab2d-9a83-4867-b43d-33e529256f10.mov)

Query All Transfers is now only called when landing on the main wallet page - or when a user scrolls to the bottom of the transaction details page. In the above video, the limit is set to per 2 transactions. In the code - we default to 10 on the transaction list page and 5 on the main wallet page.  

The query all transfers network request is no longer called on every screen. 

Also included: 
Set the spark address in zustand store on wallet initiation. No more network requests for the spark address of the current wallet. 

Simple loading states for transactions list. 

Simple loading improvements for main wallet page to show the default element instead of a blank page.  

Hide the scrollbar on transaction details page.