Fix running SO with TLS disabled. (#861)
Our Rust gRPC framework (Tonic) isn't allowing us to skip TLS
verification, or use the certs we generate, so we need to run the SOs
without TLS enabled.

For the server itself, if TLS is disabled run gRPC through a raw
net.Listen rather than http.Server, because HTTP implies HTTP/2,
which practically for Go's standard library means TLS is required.

For the clients, update the SO configuration so that cert_path is
optional, and if not specified for the SOs, create a client without TLS.
To support this, updated run-everything.sh so that when
--disable-tls is specified, a configuration is generated that doesn't
have cert-path.