Add loop waiting for deposit leaf to confirm (#880)
Occasionally we will try doing a transfer in a test, and leaves won't be available, and instead their status is "CREATING". Right now we just wait 100ms after confirming on-chain for the node to pick up the tx, but this can be flakey based on different environments. Here we add a loop that times out after 10s, that waits for the leaf to become available before moving forward.