this change fixes an invalid mnemonic error for clients consuming the spark sdk that are passing back the seed in hex string format (#914)
lrc20 wallet initialization does not account for seed being passed in string format, nor does it currently account for the seed being. of type string when it is passed into the function. 

Validate that the mnemonic is actually a valid mnemonic before we decide it is a mnemonic. 
Convert the seed to uint8 array if we receive the seed as a string.