Reverse coop exit txid in SO (#1142)
Too many times now various SDKs have provided the reversed txid. At some point you have to do some reflection and question if you're the problem.

For the last time: txid is the reverse of tx hash! If we ask the user for the txid, we should expect they are giving us the txid, not the txhash.