Allow claiming transfers after expiry (#1161)
* Add expiry parameter to coop exit, refactor test construction

* Refactor user setup for coop exit

* Add cancel coop exit test (that fails)

* Don't prevent claiming after expiry

We added expiry to allow the sender to claw back funds, if they were
doing a conditional transfer, and the receiver did not fulfill the
condition (i.e. they didn't send the lightning and get the preimage, or
they didn't broadcast/confirm the exit tx).

However, after expiry, if the receive does fulfill the condition, they
should still be able to claim it. This commit removes the condition to
fail claiming based on expiry. We don't add extra checks because we
already express whether the condition has been fulfilled by
differentiating between the pending_key_tweak and the key_tweaked
status, so if the key is fully tweaked, we know the receiver should get
the funds.

* fmt

* Separate into separate test cases