Changed base_txid / vout to be unique in `Tree`. (#1202)
At long last. Previously attempted in dac0aa62. It turns out, we had
1400 trees that were violating the uniqueness constraint for base_txid,
so that attempt needed to be rolled back.

This unique index is needed to prevent double spending, wherein a user
can create two trees spending the same UTXO.

This time, create a proper migration that deletes all trees that violate
this constraint (in regtest). Only do this in regtest because regtest
isn't real money. I have verified all of the affected trees in dev are
regtest. If this happens to affect a mainnet tree in prod, we will deal
with it manually.

Contributes to LPT-269.