Fixes local-test "used" keyshares (#1245)
The actual problem was that for whatever reason if you don't wait for
a little bit after our SOs come up and ready, the DKG might not make
it to all of the servers. So that meant that sometimes some of the servers might actually just have 0 keys while the others had some. 

This is a bit hacky, but it shouldn't happen
much/at all in production or in CI.