Generate static deposit address
Create a new API to generate a static deposit address, that allows a user to
deposit to it multile times. SE will have to mark such addresses as static and
execute a separate flow to claim a deposit that involves SSP.

Using the same API as for express deposit address, because it is 99% identical.