Ignore available tree nodes in chain watcher. (#1262)
We currently have a tree node that was marked available in
`FinalizeNodeSignatures` and subsequently split, so when chain watcher
is processing the block with the confirmed transaction it's trying to
mark it available as well.

For now, just skip that tree node to unblock our deployment. I'll follow
up with Alec tomorrow to discuss the real solution.