Setup minimum viable error code variants (#1248)
* Setup minimum viable error code variants

We want to be able to specify the type of error code specific errors
will return to the user at the place where the error occurs. This adds
just one variant, and makes the changes in our basic error
wrappers/middleware to be able to use that status.

This also adds a simple test case for input validation to make sure this
works. Proliferation of usage will come in follow up PRs if this
approach makes sense.

Move to errors package

* Doc link, stricter generic, simplify status

* fmt