Fixed validation for token revocation secrets. (#1285)
When converting `big.Int` to `[]byte` with `x.Bytes()`, the resulting
byte array will only be as long as needed to capture the integer. So if
the `big.Int` happens to be representing a private key where the leading
byte is 0, we will get a `[]byte` with length 31 rather than length 32,
which fails our `ValidateRevocationKeys` check.

Did a few things here:
- Added some helpers for safely converting between from `big.Int` to
  `secp256k1.PrivateKey` and `[]byte` to `secp256k1.PrivateKey`.
- Changed input type of `ValidateRevocationKeys` from `[][]byte` to
  `[]*secp256k1.PrivateKey` to take advantage of the type system.
- Fixed conversion from `big.Int` to `secp256k1.PrivateKey` in
  `wallet.FinalizeTokenTransaction`.