Fixed error handling in `GetUnusedSigningKeyshares`. (#1286)
If we end up having to rollback we return `tx.Rollback()`, which _can_
return an error but if it doesn't we end up squashing the error that
should be returned.

Furthermore, make sure callers are checking their results and getting
back the right number of keyshares. In `token_transaction_handler` we
were panicking when we did not get the right number back (due to the
aforementioned error squashing).