Fix query all transfers with basic identity filter (#1302)
* Fix query all transfers with basic identity filter

We had a regression in https://github.com/lightsparkdev/spark/pull/1220/files
where we were previously returning transfers for both sender and
receiver for the query_all_transfers RPC, but changed it to only
be receiver transfers. Here we add an option in `TransferFilter`
to allow a user to ask for this previous behavior.

* rename

* rename changes

* Change in js too