Add gRPC health probe to Dockerfile. (#1351)
The built-in k8s gRPC health probe doesn't support TLS. Right now we are
just doing basic TCP probes, but that leaves a bunch of TLS handshake
errors in our logs. So try using this other program that implements the
gRPC health checking protocol AND supports TLS. This PR adds that
program to our Dockerfile so we can set up an exec probe that uses it in
the SO container.