Extend timeout and Clean Coop Exit Test (#1395)
Test appears to be flaking because of a timeout in waiting for `createNewTree` to return a node. The current timeout set for this operation is 10 seconds, I believe this test is flaking due to this operation taking longer on certain iterations. Bumping the timeout to 30 seconds should eliminate this flake. I'm also changing the error handling to use `require` instead of `assert` as it provides better error messaging and is congruent with how we are handling errors in other integration tests.