[so] Create ent Utxo
Ent Utxo will be used track UTXOs deposited to user's static deposit addresses.
If an SSP submits a claim for an UTXO, all SOs will validated that the claim is
for a previously seen UTXO and then assign this UTXO to the SSP, so other SSP
won't be able to claim it.

Towards LPT-314